Embiggen VQAHandle
long long pointer

+++ CODE/conquer.cpp
@@ -2661,7 +2661,7 @@ long MixFileHandler(VQAHandle * vqa, long action, void * buffer, long nbytes)
 				error = file->Open((char *)buffer, READ);
 
 				if (error != -1) {
-					vqa->VQAio = (unsigned long)file;
+					vqa->VQAio = (uintptr_t)file;
 					error = 0;
 				} else {
 					delete file;
